Seperated the http requests into different folders
